chore[eventstore/mongov2]: move snapshot logic to separate file
